Moving test/unittests/{env,testutil}.{h,cc} to test/common/
