+++ test/unittests/main.cc
@@ -6,7 +6,7 @@
 
 #include <cassert>
 
-#include "env.h"
+#include "../common/env.h"
 #include "monitor.h"
 
 int main(int argc, char **argv) {
