+++ test/unittests/t_catalog.cc
@@ -11,7 +11,7 @@
 #include "catalog_rw.h"
 #include "hash.h"
 #include "shortstring.h"
-#include "testutil.h"
+#include "../common/testutil.h"
 
 using namespace std;  // NOLINT
