+++ test/unittests/t_catalog_counters.cc
@@ -4,7 +4,7 @@
 
 #include <gtest/gtest.h>
 
-#include "testutil.h"
+#include "../common/testutil.h"
 
 #include "catalog_counters.h"
