+++ test/unittests/t_catalog_traversal.cc
@@ -12,7 +12,7 @@
 #include "hash.h"
 #include "manifest.h"
 #include "prng.h"
-#include "testutil.h"
+#include "../common/testutil.h"
 
 using swissknife::CatalogTraversal;
