+++ test/unittests/t_garbage_collector.cc
@@ -14,7 +14,7 @@
 #include "hash.h"
 #include "manifest.h"
 #include "prng.h"
-#include "testutil.h"
+#include "../common/testutil.h"
 
 using swissknife::CatalogTraversal;
 using upload::SpoolerDefinition;
