+++ test/unittests/t_test_utils.cc
@@ -4,7 +4,7 @@
 
 #include <gtest/gtest.h>
 
-#include "testutil.h"
+#include "../common/testutil.h"
 
 // This is very meta!
 // We are testing the utility functions used for testing...
